generator/pyproject.toml
[project]
name = "generator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13.0"
dependencies = [
    "gpxpy>=1.6.2",
    "lxml>=6.0.0",
    "python-magic>=0.4.27",
]
